requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
